Actualizar las funciones en get_next_line.h y get_next_line_utils.c: agregar declaraciones de funciones y cambiar el tipo de retorno de ft_strlen a size_t

+++ get_next_line_utils.c
@@ -30,9 +30,9 @@ char	*ft_strdup(const char *str)
 	return (str2);
 }
 
-int	ft_strlen(const char *str)
+size_t	ft_strlen(const char *str)
 {
-	int	length;
+	size_t	length;
 
 	length = 0;
 	while (*str)
